// src/app/core/models/orcamento.model.ts
export interface OrcamentoRequest {
  tipoVaga: string;
  dataEntrada: string;
  qtdDias: number;
}

export interface OrcamentoResponse {
  tipoVaga: string;
  dataEntrada: string;
  qtdDias: number;
  dataSaidaPrevista: string;
  valorDiaria: number;
  valorTotalCartao: number;
  valorTotalPixDinheiro: number;
  descontoPixDinheiroPorDia: number;
  economiaTotal: number;
  vagasDisponiveis: boolean;
  vagasRestantes: number;
}